Fix script not working if optional cli-arg "maxprice" not given
Considering the ‘maxprice’ arg is optional, the script doesn’t work(returns 400 error) when maxprice is not given. This is because the API url ends up having “None” in the last parameter. This is a fix for that, which will assign maxprice with an empty string when arg is not given